// src/app/models/votacion.ts
export class Votacion {
    "id": string;
    "candidato": string;
    "votos": number;
    "designadoPrincipal": string;
    "partido": string;
    "color": string;
}
export class VotacionWithoutID {
    "id": string;
    "candidato": string;
    "votos": number;
    "designadoPrincipal": string;
    "partido": string;
    "color": string;
}